refactor: extract desktop update progress
